modify package structure
