+++ .gitignore
@@ -128,6 +128,3 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-# backend-app
-backend-app/.idea/
-
