+++ backend-app/.gitignore
@@ -1,0 +1,5 @@
+# IDE
+.idea/
+
+# Environment
+.env
