+++ backend-app/app/config.py
@@ -1,0 +1,16 @@
+from pydantic import BaseSettings
+
+
+class Settings(BaseSettings):
+    db_engine: str = "postgresql"
+    db_user: str
+    db_password: str
+    db_host: str = "localhost"
+    db_port: int = 54321
+    db_name: str = "murmur"
+
+    class Config:
+        env_file = ".env"
+
+
+settings = Settings()
