+++ backend-app/app/middleware/security.py
@@ -1,0 +1,6 @@
+from fastapi.security import OAuth2PasswordBearer
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/token")
+origins = [
+    "http://localhost:3000"
+]
