+++ backend-app/requirements.txt
@@ -4,6 +4,7 @@ h11==0.9.0
 httptools==0.1.1
 psycopg2-binary==2.8.5
 pydantic==1.5.1
+python-dotenv==0.13.0
 python-multipart==0.0.5
 six==1.14.0
 SQLAlchemy==1.3.16
