+++ docs/design/dev_ops_design.md
@@ -28,6 +28,6 @@
 - フロントエンド
   - ローカルの 3000 ポートで起動
 - バックエンド
-  - ローカルからコンテナへ 8080:80 でルーティング
+  - ローカルの 8000 ポートで起動
 - データベース
   - ローカルからコンテナへ 5432:5432 でルーティング
